// src/api/gen/model/v1alpha1ApplicationStatus.ts
/**
 * Consolidate Services
 * Description of all APIs
 *
 * The version of the OpenAPI document: version not set
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';
import { V1Time } from './v1Time';
import { V1alpha1ApplicationCondition } from './v1alpha1ApplicationCondition';
import { V1alpha1ApplicationSummary } from './v1alpha1ApplicationSummary';
import { V1alpha1HealthStatus } from './v1alpha1HealthStatus';
import { V1alpha1OperationState } from './v1alpha1OperationState';
import { V1alpha1ResourceStatus } from './v1alpha1ResourceStatus';
import { V1alpha1RevisionHistory } from './v1alpha1RevisionHistory';
import { V1alpha1SyncStatus } from './v1alpha1SyncStatus';

export class V1alpha1ApplicationStatus {
    'conditions'?: Array<V1alpha1ApplicationCondition>;
    'health'?: V1alpha1HealthStatus;
    'history'?: Array<V1alpha1RevisionHistory>;
    'observedAt'?: V1Time;
    'operationState'?: V1alpha1OperationState;
    'reconciledAt'?: V1Time;
    'resources'?: Array<V1alpha1ResourceStatus>;
    'sourceType'?: string;
    'summary'?: V1alpha1ApplicationSummary;
    'sync'?: V1alpha1SyncStatus;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "conditions",
            "baseName": "conditions",
            "type": "Array<V1alpha1ApplicationCondition>"
        },
        {
            "name": "health",
            "baseName": "health",
            "type": "V1alpha1HealthStatus"
        },
        {
            "name": "history",
            "baseName": "history",
            "type": "Array<V1alpha1RevisionHistory>"
        },
        {
            "name": "observedAt",
            "baseName": "observedAt",
            "type": "V1Time"
        },
        {
            "name": "operationState",
            "baseName": "operationState",
            "type": "V1alpha1OperationState"
        },
        {
            "name": "reconciledAt",
            "baseName": "reconciledAt",
            "type": "V1Time"
        },
        {
            "name": "resources",
            "baseName": "resources",
            "type": "Array<V1alpha1ResourceStatus>"
        },
        {
            "name": "sourceType",
            "baseName": "sourceType",
            "type": "string"
        },
        {
            "name": "summary",
            "baseName": "summary",
            "type": "V1alpha1ApplicationSummary"
        },
        {
            "name": "sync",
            "baseName": "sync",
            "type": "V1alpha1SyncStatus"
        }    ];

    static getAttributeTypeMap() {
        return V1alpha1ApplicationStatus.attributeTypeMap;
    }
}